+++ src/Sulu/Bundle/Sales/CoreBundle/Resources/public/js/components/item-table/main.js
@@ -92,6 +92,12 @@ define([
             urlFilter: {}
         },
 
+        TYPES = {
+            PRODUCT: 0,
+            CUSTOM: 1,
+            ADDON: 2
+        },
+
         urls = {
             products: '/admin/api/products{?filter*}',
             product: '/admin/api/products/',
@@ -147,7 +153,8 @@ define([
             currency: Config.get('sulu_sales_core').default_currency,
             useProductsPrice: false,
             tax: 0,
-            supplierName: ''
+            supplierName: '',
+            type: TYPES.PRODUCT
         },
 
         templates = {
@@ -1443,6 +1450,9 @@ define([
                 updateGlobalPrice.call(this, rowId);
                 refreshItemsData.call(this);
             } else {
+                // Given product has to be of type custom.
+                data.type = TYPES.CUSTOM;
+
                 // Add default for using products price.
                 if (typeof data.useProductsPrice === 'undefined') {
                     data.useProductsPrice = false;