+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 CHANGELOG for Sulu Sales
 ========================
 
-* dev-develop    
+* dev-develop
 
+    * BUGFIX      [CoreBundle]    Fixed creating independent items in item table caused by 
+                                  changes in pricing bundle 
     * ENHANCEMENT [General]       Moved db diagram from sulu/docs to sales
 
 * 0.4.0 (2016-07-06)